Clean up R_InitTables and do a quick check against a memory dump obtained from pcem.
It seems accurate, but I need to check thoroughly, and I'm not sure if pcem is the best comparison source, but I did test with both an emulated FPU and no emulated FPU, using watcom's emulation instead.

+++ src/hmaph.c
@@ -45,7 +45,7 @@ void R_MapRowHigh(void)
 
 	local_8 = mr_x1;
 	iVar1 = &colormaps[mr_colormap * 256];
-	local_14 = screenbuffer + (int)(planewidthlookup[mr_y + windowy] + (mr_x1 + windowx >> 2));
+	local_14 = screenbuffer + planewidthlookup[mr_y + windowy] + (mr_x1 + windowx >> 2);
 	local_10 = ((mr_x1 + windowx) & 3U) * 16000;
 
 	do
